+++ ionoscloud/resource_kafka_cluster.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"log"
+	"slices"
 	"strings"
 	"time"
 
@@ -178,6 +179,11 @@ func resourceKafkaClusterImport(ctx context.Context, d *schema.ResourceData, met
 	parts := strings.Split(d.Id(), ":")
 	if len(parts) != 2 {
 		return nil, fmt.Errorf("expected ID in the format location:cluster_id")
+	}
+
+	location := parts[0]
+	if !slices.Contains(kafka.AvailableLocations, location) {
+		return nil, fmt.Errorf("invalid location: %v, location must be one of: %v", location, kafka.AvailableLocations)
 	}
 
 	if err := d.Set("location", parts[0]); err != nil {